+++ utilities/NekMesh/InputModules/InputGmsh.cpp
@@ -868,7 +868,7 @@ void InputGmsh::Process()
             }
         }
     }
-    m_mshFile.close();
+    m_mshFile.reset();
 
     // Go through element and remap tags if necessary.
     map<int, map<LibUtilities::ShapeType, int> > compMap;